+++ include/roq/fbs/encode.hpp
@@ -360,8 +360,7 @@ auto encode(B &builder, roq::MarketByPriceUpdate const &value) {
       value.exchange_sequence,
       encode(builder, value.price_decimals),
       encode(builder, value.quantity_decimals),
-      value.max_depth,
-      value.checksum);
+      value.max_depth);
 }
 
 template <typename B>